Max upload limit (#90)
* File: Add maximum upload per user

A maximum upload in bytes per user can be set in the config file. For
this to work, the number of bytes uploaded per user is tracked in the
database and checked when uploading a new file.

The new file being uploaded is only rejected _after_ it has been saved
since that is the easiest way to get the size of the file that was
uploaded. The file is then deleted.

A column in the database for the table `user` has been added called
`uploaded_bytes` to track the number of bytes uploaded per user.
This can not be decreased again as the files never get deleted on the
server but are instead marked as deleted.

* remove unnecessary declaration of `_file` in api

* File: extend calls with `fileSize` parameter

Extend the parameters of `SaveFile` (and functions calling it) to
include a `fileSize` parameter to avoid saving the file before being
able to check whether the user has hit their upload limit.

* Add note for `usersx` route in main.go

* Files: upload limit of 0 disables limit

* Generate new sqlboiler models