Use Middleware to make id's available
User and Role IDs are now available through the Cookie instead of
having to get them through boilerplate code.
To make this work, every API function that needs the Cookie _has to_ be
added to the router in `auth`, which groups all API functions that need
authorization via the cookie.