Fix wrong validation of deadlines in exam
Deadlines were converted to `null.Time` and checked with `isZero`, but
this only checks whether the `null.Time` is valid or not. Creating it
with `null.TimeFrom()` will always be valid so it it replaced with
`null.NewTime()` which takes an extra parameter for the validity, which
is checked with the `isZero` method on `time`.